Changing votingpower's data type from float64 to uint64 (#1744)
* Change data type of voting power (float64 --> uint64

* Fixed weighted validators test

+++ consensus/istanbul/validator.go
@@ -34,7 +34,7 @@ type Validator interface {
 	String() string
 
 	RewardAddress() common.Address
-	VotingPower() float64
+	VotingPower() uint64
 	Weight() int
 }
 
@@ -101,7 +101,7 @@ type ValidatorSet interface {
 
 	Proposers() []Validator // TODO-Klaytn-Issue1166 For debugging
 
-	TotalVotingPower() float64
+	TotalVotingPower() uint64
 }
 
 // ----------------------------------------------------------------------------
